feat: get scanID
